#660 - store - split category module
