+++ src/store/modules/category/actions.js
@@ -1,25 +1,9 @@
-import * as types from '../mutation-types'
-import { entityKeyName } from '../../lib/entities'
-import { slugify } from '../../helpers'
+import * as types from 'src/store/mutation-types'
+import { quickSearchByQuery } from 'src/api/search'
+import { entityKeyName } from 'src/lib/entities'
 const bodybuilder = require('bodybuilder')
-import { quickSearchByQuery } from '../../api/search'
 
-const state = {
-  list: [],
-  current: {},
-  filters: { color: [], size: [], price: [] },
-  breadcrumbs: {routes: []},
-  current_path: [] // list of categories from root to current
-}
-
-const getters = {
-  current: (state) => state.current,
-  list: (state) => state.list
-}
-
-// actions
-const actions = {
-
+export default {
   /**
    * Reset current category and path
    * @param {Object} context
@@ -55,7 +39,6 @@ const actions = {
       console.error(err)
     })
   },
-
   /**
    * Load category object by specific field - using local storage/indexed Db
    * loadCategories() should be called at first!
@@ -130,48 +113,4 @@ const actions = {
       }
     })
   }
-}
-
-// mutations
-const mutations = {
-
-  [types.CATEGORY_UPD_CURRENT_CATEGORY] (state, category) {
-    state.current = category
-  },
-  [types.CATEGORY_UPD_CURRENT_CATEGORY_PATH] (state, path) {
-    state.current_path = path // TODO: store to cache
-  },
-
-  [types.CATEGORY_UPD_CATEGORIES] (state, categories) {
-    state.list = categories.items
-
-    for (let category of state.list) {
-      let catSlugSetter = (category) => {
-        for (let subcat of category.children_data) { // TODO: fixme and move slug setting to vue-storefront-api
-          subcat = Object.assign(subcat, { slug: subcat.hasOwnProperty('name') ? slugify(subcat.name) + '-' + subcat.id : '' })
-          catSlugSetter(subcat)
-        }
-      }
-      catSlugSetter(category)
-      const catCollection = global.db.categoriesCollection
-      try {
-        catCollection.setItem(entityKeyName('slug', category.slug.toLowerCase()), category).catch((reason) => {
-          console.error(reason) // it doesn't work on SSR
-        }) // populate cache by slug
-        catCollection.setItem(entityKeyName('id', category.id), category).catch((reason) => {
-          console.error(reason) // it doesn't work on SSR
-        }) // populate cache by id
-      } catch (e) {
-        console.error(e)
-      }
-    }
-  }
-}
-
-export default {
-  namespaced: true,
-  state,
-  getters,
-  actions,
-  mutations
 }
