+++ src/store/modules/category/getters.js
@@ -1,0 +1,4 @@
+export default {
+  current: (state) => state.current,
+  list: (state) => state.list
+}
